requirements.txt
Pillow==11.1.0
termcolor==2.5.0
